+++ test/runner/BUILD
@@ -26,6 +26,7 @@ go_binary(
         "//test/runner/gtest",
         "//test/trace/config",
         "//test/uds",
+        "@com_github_kr_pty//:go_default_library",
         "@com_github_moby_sys_capability//:go_default_library",
         "@com_github_opencontainers_runtime_spec//specs-go:go_default_library",
         "@org_golang_x_sys//unix:go_default_library",
